Update tictactoe example.
