+++ .gitignore
@@ -1,4 +1,6 @@
-build/*
+build/
 node_modules
 .DS_Store
 ledger.json
+package-lock.json
+yarn.lock
